refactor(terminal): 拆出主题壁纸与设置事件监听 hook
- 新增 components/terminal/：useTerminalTheme（主题对象缓存/壁纸状态/xterm 主题与容器 CSS 变量同步）、useTerminalSettingsEvents（快捷键/回显/时间戳/命令块/字体/鼠标手势/关键字高亮等设置变更事件同步）
- Terminal.tsx 3568 -> 3391 行
- 2 处存量 hex 字面量随主题同步逻辑迁入 .ts（styles:check 依机制自动下调基线 36 -> 34）
