+++ frontend/scripts/style-baseline.json
@@ -1,5 +1,5 @@
 {
-  "hexColor": 36,
+  "hexColor": 34,
   "numericFontSize": 20,
   "magicZIndex": 4
 }
